Fix a warning.
MFC after:	1 week


git-svn-id: svn+ssh://svn.freebsd.org/base/head@323776 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/netinet/sctp_input.c
@@ -2579,7 +2579,7 @@ sctp_handle_cookie_echo(struct mbuf *m, int iphlen, int offset,
 		} else {
 			staleness = diff.tv_sec * 1000000;
 		}
-		if (UINT32_MAX - staleness >= diff.tv_usec) {
+		if (UINT32_MAX - staleness >= (uint32_t)diff.tv_usec) {
 			staleness += diff.tv_usec;
 		} else {
 			staleness = UINT32_MAX;
